credit-pricing-tool/requirements.txt
pyyaml>=6.0
pandas>=2.0
pyarrow>=14.0
docling>=2.0
docling-core>=2.0
ocrmypdf>=16.0
anthropic>=0.40.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
httpx>=0.25.0
supabase>=2.0
python-dotenv>=1.0
